scaled-cpikan-dhm/requirements.txt
# Basic requirements
numpy
torch
pyyaml

# For visualization
matplotlib
